Add option to use custom loaders
Nunjucks allows for custom loaders for more complex loading, like from
a database. jstransformer-nunjucks doesn't support it, because it uses
the Nunjucks simple API by default. This commit changes the default
behavior to switch to the Nunjucks Environment constructor if `loaders`
is available as a key in the `options` dictionary object.